fix some settings
* fix setup script retry bug.
* fix Multibyte Space Syntax.

+++ setup.sh
@@ -8,17 +8,27 @@ do
 done
 
 # setup git submodule plugins
+echo "Update git Submodules."
 cd $DIR
 git pull
 git submodule init
 git submodule update
 
 # setup my vimrc files
-mkdir -p $DIR/.vim/bundle
-cd $DIR/.vim/bundle
-ln -s ../vimrc
+BUNDLE_DIR=$DIR/.vim/bundle
+if [ ! -d "$BUNDLE_DIR" ]; then
+    echo "Create .vim/bundle dir."
+    mkdir -p $BUNDLE_DIR
+fi
+if [ ! -h "$BUNDLE_DIR/vimrc" ]; then
+    echo "Symlink vimrc dir."
+    cd $BUNDLE_DIR
+    ln -s ../vimrc
+fi
 
-# setup oh-my-zsh theme
+# setup oh-my-zsh theme bugfix
 cd $DIR/.oh-my-zsh/themes
-sed -e "s/\(.*foopath.*\)/#\1/" linuxonly >linuxonly.zsh-theme
-
+if [ ! -f $DIR/.oh-my-zsh/themes/linuxonly.zsh-theme ]; then
+    echo "Fix oh-my-zsh's linuxonly theme And Craete linuxonly.zsh-theme."
+    sed -e "s/\(.*foopath.*\)/#\1/" linuxonly >linuxonly.zsh-theme
+fi
